requirements.txt
asgiref==3.5.0
astroid==2.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Distance==0.1.3
distlib==0.3.4
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.4.2
GDAL @ file:///C:/Users/Admin/Desktop/saurav/beachplus_project/updated%20beachplus_custom%20user/GDAL-3.4.1-cp310-cp310-win_amd64.whl
geographiclib==1.52
geopy==2.2.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
microsoft==0.1.2
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.4.1
pony==0.7.16
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
store==2020.2.10
toml==0.10.2
twilio==7.7.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
wrapt==1.13.3